requirements.txt
appdirs==1.4.4
apps==0.0.0.0.dev20200918
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
distlib==0.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-suit==0.2.28
filelock==3.0.12
form==0.0.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.3
Pillow==8.0.1
psycopg2==2.8.6
pylint==2.6.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
view==0.1
virtualenv==20.2.2
wrapt==1.12.1
